requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
sqlalchemy
aiosqlite
